BACKEND/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
psycopg2
SQLALchemy  
Alembic
passlib[bcrypt]
python-jose[cryptography]